requirements.txt
Flask
mysql-connector-python
paramiko
python-dotenv
requests